[01/30/20-02/05/20] Winter 2020, 2020-02-06T15:38:13+00:00, version D
1. We only considered "shuttle trips" to be those trips which weren't on the
original route. But, for shuttles where we're truncating part of a route
along with a shuttle, we do need to keep those other trips. We change the
logic instead to treat any trip which isn't the one we're currently
processing as a "shuttle trip".
2. We weren't copying the shapes when running multiple disruptions. Then, we
weren't processing the shapes in an order which allowed for a truncation to
be applied to an already truncated shape.

https://cdn.mbtace.com/archive/20200130.zip