[05/26/20-05/27/20] Spring 2020, 2020-06-02T22:28:51+00:00, version D
<https://cdn.mbtace.com/archive/20200526.zip>

Previously we were inconsistent about whether there was an `__init__.py`
file in each test directory or not, which impacts how pytest loads tests
in that directory (with `__init__` they are loaded as fully-qualified
"packages", without they are loaded as top-level modules). This had the
odd consequence that using the `pytest` script (vs. `python -m pytest`)
to run a specific test file would work for some files, but crash with a
`ModuleNotFoundError` on others.
Ref: https://docs.pytest.org/en/latest/pythonpath.html