[04/05/21-04/07/21] Spring 2021, 2021-04-12T19:51:14+00:00, version D
<https://cdn.mbtace.com/archive/20210405.zip>

This enhances `RouteTypeDisruption` (now `ReplaceService`) to enable
directly specifying lists of service IDs to deactivate and activate,
instead of looking up services based on dates and route criteria. This
allows activating services that don't exist at all in the pre-disruption
calendar, which will be necessary to activate the "storm schedules" (and
potentially other schedules) under the new HASTUS import process.