[11/26/21-11/28/21] Fall 2021, 2021-12-03T20:56:22+00:00, version D
<https://cdn.mbtace.com/archive/20211126.zip>

This was used to enable secrets in the tests, but had some odd side
effects as GitHub assumes only one branch at a time can be "deployed"
to an environment. We now have the required variables set as repository
secrets, so this workaround shouldn't be needed.
Since Actions throws strange errors when a "deploy workflow" is changed
to a "non-deploy workflow", this also changes the workflow name, making
it unrelated to the old one as far as Actions is concerned.